moved all endpoints into keys
