+++ .prettierignore
@@ -11,3 +11,4 @@ services/*/apps/api/src/metadata.ts
 services/*/docs/index.html
 **/metadata.ts
 openapi-specs/*
+jest.config.json
