+++ apps/account-api/src/controllers/v1/index.ts
@@ -1,5 +1,4 @@
 export * from './accounts-v1.controller';
 export * from './delegation-v1.controller';
-export * from './graphs-v1.controller';
 export * from './handles-v1.controller';
 export * from './keys-v1.controller';
