+++ apps/account-api/src/services/index.ts
@@ -1,5 +1,4 @@
 export * from './accounts.service';
 export * from './delegation.service';
-export * from './graphs.service';
 export * from './handles.service';
 export * from './keys.service';
