+++ libs/account-lib/src/types/enums.ts
@@ -5,5 +5,5 @@ export enum TransactionType {
   SIWF_SIGNUP = 'SIWF_SIGNUP',
   SIWF_SIGNIN = 'SIWF_SIGNIN',
   ADD_KEY = 'ADD_KEY',
-  ADD_GRAPH_KEY = 'ADD_GRAPH_KEY',
+  ADD_PUBLIC_KEY_AGREEMENT = 'ADD_PUBLIC_KEY_AGREEMENT',
 }
