+++ openapi-specs/content-publishing.openapi.json
@@ -322,6 +322,89 @@
           "health"
         ]
       }
+    },
+    "/dev/request/{jobId}": {
+      "get": {
+        "operationId": "DevelopmentControllerV1_requestJob",
+        "summary": "Get a Job given a jobId",
+        "description": "ONLY enabled when ENVIRONMENT=\"dev\".",
+        "parameters": [
+          {
+            "name": "jobId",
+            "required": true,
+            "in": "path",
+            "schema": {
+              "type": "string"
+            }
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": ""
+          }
+        },
+        "tags": [
+          "dev"
+        ]
+      }
+    },
+    "/dev/asset/{assetId}": {
+      "get": {
+        "operationId": "DevelopmentControllerV1_getAsset",
+        "summary": "Get an Asset given an assetId",
+        "description": "ONLY enabled when ENVIRONMENT=\"dev\".",
+        "parameters": [
+          {
+            "name": "assetId",
+            "required": true,
+            "in": "path",
+            "schema": {
+              "type": "string"
+            }
+          }
+        ],
+        "responses": {
+          "2XX": {
+            "content": {
+              "application/octet-stream": {
+                "schema": {
+                  "type": "string",
+                  "format": "binary"
+                }
+              }
+            },
+            "description": ""
+          }
+        },
+        "tags": [
+          "dev"
+        ]
+      }
+    },
+    "/dev/dummy/announcement/{queueType}/{count}": {
+      "post": {
+        "operationId": "DevelopmentControllerV1_populate",
+        "summary": "Create dummy announcement data",
+        "description": "ONLY enabled when ENVIRONMENT=\"dev\".",
+        "parameters": [
+          {
+            "name": "count",
+            "required": true,
+            "in": "path",
+            "schema": {
+              "type": "number"
+            }
+          }
+        ],
+        "responses": {
+          "201": {
+            "description": ""
+          }
+        },
+        "tags": [
+          "dev"
+        ]
+      }
     }
   },
   "info": {
